Separate RPC and Admin server test
